feat: Handle API being down/blocked, instead of making the site hang until the configured timeout happens.
We now have a configured timeout for the API client as the following:
- Connection timeout: 0.5 seconds at most
- Response timeout: 5 seconds at most (yes, if the API is slow, we will consider it down)

And in the plugin, we have also implemented a check that has a timeout of 0.25 seconds (but always goes for 1 second anyway for some reason), and if we decide that the API is not responding fast enough, a warning saying that the plugin cannot connect to the API will be shown.

In the admin pages, an error will be shown, with the suggestion to contact the EduAdmin Support for help.

Fixed another possible injection in template variables in course list view.

fixes #153